+++ hardhat-config/providers.config.js
@@ -3,26 +3,26 @@ module.exports = {
     infura: {
       mainnet: '{{PROVIDER_URL_INFURA_MAINNET}}',
       goerli: '{{PROVIDER_URL_INFURA_GOERLI}}',
-      polygon: '{{PROVIDER_URL_INFURA_POLYGON}}',
-      polygonMumbai: '{{PROVIDER_URL_INFURA_POLYGONMUMBAI}}',
+      matic: '{{PROVIDER_URL_INFURA_MATIC}}',
+      mumbai: '{{PROVIDER_URL_INFURA_MUMBAI}}',
     },
     alchemy: {
       mainnet: '{{PROVIDER_URL_ALCHEMY_MAINNET}}',
       goerli: '{{PROVIDER_URL_ALCHEMY_GOERLI}}',
-      polygon: '{{PROVIDER_URL_ALCHEMY_POLYGON}}',
-      polygonMumbai: '{{PROVIDER_URL_ALCHEMY_POLYGONMUMBAI}}',
+      matic: '{{PROVIDER_URL_ALCHEMY_MATIC}}',
+      mumbai: '{{PROVIDER_URL_ALCHEMY_MUMBAI}}',
     },
     maticvigil: {
-      polygon: '{{PROVIDER_URL_MATICVIGIL_POLYGON}}',
-      polygonMumbai: '{{PROVIDER_URL_MATICVIGIL_POLYGONMUMBAI}}',
+      matic: '{{PROVIDER_URL_MATICVIGIL_MATIC}}',
+      mumbai: '{{PROVIDER_URL_MATICVIGIL_MUMBAI}}',
     },
     binance: {
       bsc: 'https://bsc-dataseed.binance.org/',
-      bscTestnet: 'https://data-seed-prebsc-1-s1.binance.org:8545/',
+      bsctest: 'https://data-seed-prebsc-1-s1.binance.org:8545/',
     },
     exchain: {
       okex: 'https://exchainrpc.okex.org',
-      okexTestnet: 'https://exchaintestrpc.okex.org',
+      okextest: 'https://exchaintestrpc.okex.org',
     },
   },
 };