+++ hardhat-plugins/node-providers/index.js
@@ -8,12 +8,22 @@ extendConfig((config, userConfig) => {
 
   for (const networkName of Object.keys(networksConfig)) {
     const network = networksConfig[networkName];
-    const networkProvider = provider || network.defaultProvider;
-    if (networkProvider !== undefined) {
-      if (providersConfig[networkProvider] !== undefined && providersConfig[networkProvider][network.url] !== undefined) {
-        network.url = providersConfig[networkProvider][network.url].replace(/.*(\{\{.+\}\}).*/, function (match, p) {
+    if (network.live) {
+      let networkURL;
+      if (providersConfig[provider] && providersConfig[provider][networkName]) {
+        networkURL = providersConfig[provider][networkName];
+      } else if (providersConfig[network.defaultProvider] && providersConfig[network.defaultProvider][networkName]) {
+        networkURL = providersConfig[network.defaultProvider][networkName];
+      }
+      if (networkURL) {
+        network.url = networkURL.replace(/.*(\{\{.+\}\}).*/, function (match, p) {
           return match.replace(p, process.env[p.replace(/\{|\}/g, '')]);
         });
+      } else {
+        console.log(
+          `Could not retrieve a provider URL for network ${networkName}. ` +
+            `env.NODE_PROVIDER=${provider} network.defaultProvider=${network.defaultProvider}`
+        );
       }
     }
   }